Renamed sidenav folder to ResponsiveDrawer.
Added isLoading variable from useAuth0() hook.

Chat.js Fixed path to accomodate new folder name for ResponsiveDrawer.  Added axios import.  Added state variable and method [userInDb, setUserInDb] = useState as a boolean to determine whether to add user to DB.  Added console logs to debug and develop.  Added logic to call a crudRoute that creates a new User record in the DB for the currently signed in user if non exists inside useEffect().

Login.js destructered isAuthenticated prop in function parameter.  Added some JSX to return statement that conditionally renders the Login button if there is no currently authenticated user, or renders a CHAT button if there is a currently authenticated user.

index.js changed the redirectUri to main landing page instead of chat page.  Removed bcrypt and email validation which is no longer necessary because we are implementing Auth0.

crudRoutes.js  Removed email validation from the /createUser route since we are not storing passwords in the DB.  Added a /checkIfUserExistsAndCreate route that adds a user to the DB if its not already in the DB.

App.js Added user, isAuthenticated, and isLoading props to <Login>